The drawing function in this assignment will be similar to the mid-term exam. In this assignment, you will develop a simple app for drawing charts. The main activity has UI elements for data input. Another activity draws the chart from data.
The UI of the main activity:

Main.png

Initially the data table has a blank header row and a data row. When button 'Add more data' is clicked, a new blank row will be added. It is put in a ScrollView so more data can be added without changing its size.

When the 'Draw bar chart' is clicked, the chart drawing activity should look like this

Barchart.png

When the 'Draw pie chart' is clicked, the chart drawing activity should look like this

Piechart.png